+++ django_email_learning/personalised/views.py
@@ -688,9 +688,6 @@ class CertificateView(BaseTemplateView):
                         ),
                         "issue_date": _("Issued on"),
                         "certificate_number": _("Certificate Number"),
-                        "organization_team": _("{organization_name} Team").format(
-                            organization_name=certificate.enrollment.course.organization.name
-                        ),
                     },
                 }
                 | self.get_app_context(),